spack.mirror.Mirror -> spack.mirrors.mirror.Mirror
