+++ lib/spack/spack/cmd/common/arguments.py
@@ -15,6 +15,7 @@ import spack.config
 import spack.deptypes as dt
 import spack.environment as ev
 import spack.mirror
+import spack.mirrors.mirror
 import spack.reporters
 import spack.spec
 import spack.store
@@ -689,7 +690,7 @@ def mirror_name_or_url(m):
 
     # If there's a \ or / in the name, it's interpreted as a path or url.
     if "/" in m or "\\" in m or m in (".", ".."):
-        return spack.mirror.Mirror(m)
+        return spack.mirrors.mirror.Mirror(m)
 
     # Otherwise, the named mirror is required to exist.
     try:
@@ -700,14 +701,14 @@ def mirror_name_or_url(m):
 
 def mirror_url(url):
     try:
-        return spack.mirror.Mirror.from_url(url)
+        return spack.mirrors.mirror.Mirror.from_url(url)
     except ValueError as e:
         raise argparse.ArgumentTypeError(str(e)) from e
 
 
 def mirror_directory(path):
     try:
-        return spack.mirror.Mirror.from_local_path(path)
+        return spack.mirrors.mirror.Mirror.from_local_path(path)
     except ValueError as e:
         raise argparse.ArgumentTypeError(str(e)) from e
